workflows/requirements.txt
azure-functions
azure-data-tables
azure-identity
azure-keyvault-secrets
msgraph-sdk
aiohttp
pydantic
PyYAML
pytest
pytest-asyncio
python-dotenv
debugpy
